Revise approach for running vehicles.
Give up trying to eliminate vehicle load order advantages. The previous
approach was getting more and more complicated, and wouldn't actually
make the outcome of a battle independendant of the order the vehicles
were loaded. This new approach just runs the vehicles and executes
actions as drcalls are called. This makes life simpler for everyone. To
allieviate the load order problem, bullets will be slow, take mutliple
steps to reach thier target, and be stored in the worldstate.